cfg80211: Add AP stopped interface
AP stopped interface can be used to indicate that the AP mode has
stopped functioning, WLAN driver may have encountered errors that has
forced the driver to stop the AP mode.

When the driver is in P2P-Go mode, and when it goes thru automatic
recovery from firmware crashes, it uses this interface to notify the
userspace that the group has been deleted.

Signed-off-by: Sameer Thalappil <sameert@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>